+++ crates/primitives/state/src/lib.rs
@@ -1,4 +1,8 @@
+//! Starknet state primitives.
 #![cfg_attr(not(feature = "std"), no_std)]
+
+#[doc(hidden)]
+extern crate alloc;
 
 use blockifier::execution::contract_class::ContractClass;
 use blockifier::state::cached_state::{ContractStorageKey, StateChangesCount};
@@ -8,6 +12,8 @@ use starknet_api::api_core::{ClassHash, CompiledClassHash, ContractAddress, Nonc
 use starknet_api::hash::StarkFelt;
 use starknet_api::state::StorageKey;
 use starknet_api::stdlib::collections::HashMap;
+
+pub mod rpc;
 
 type ContractClassMapping = HashMap<ClassHash, ContractClass>;