+++ crates/primitives/transactions/src/lib.rs
@@ -20,7 +20,7 @@ use blockifier::execution::contract_class::ContractClass;
 use blockifier::transaction::transaction_types::TransactionType;
 use derive_more::From;
 use starknet_api::transaction::Fee;
-use starknet_core::types::{TransactionExecutionStatus, TransactionFinalityStatus};
+use starknet_core::types::{MsgFromL1, TransactionExecutionStatus, TransactionFinalityStatus};
 use starknet_ff::FieldElement;
 
 const SIMULATE_TX_VERSION_OFFSET: FieldElement =
@@ -30,7 +30,6 @@ const SIMULATE_TX_VERSION_OFFSET: FieldElement =
 // pub mod utils;
 use mp_felt::Felt252Wrapper;
 
-// TODO(antiyro): remove this when released: https://github.com/xJonathanLEI/starknet-rs/blame/fec81d126c58ff3dff6cbfd4b9e714913298e54e/starknet-core/src/types/serde_impls.rs#L175
 #[derive(Debug, Clone, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct TransactionStatus {
@@ -210,4 +209,16 @@ pub struct HandleL1MessageTransaction {
     pub contract_address: Felt252Wrapper,
     pub entry_point_selector: Felt252Wrapper,
     pub calldata: Vec<Felt252Wrapper>,
+}
+
+impl From<MsgFromL1> for HandleL1MessageTransaction {
+    fn from(msg: MsgFromL1) -> Self {
+        let calldata = msg.payload.into_iter().map(|felt| felt.into()).collect();
+        Self {
+            contract_address: msg.to_address.into(),
+            nonce: 0u32.into(),
+            entry_point_selector: msg.entry_point_selector.into(),
+            calldata,
+        }
+    }
 }